[refactor] create contexts, hooks, providers directory and split alert
