+++ src/components/Product.tsx
@@ -1,5 +1,4 @@
 import styled, { css } from 'styled-components';
-import { useState } from 'react';
 import { CustomButton } from './Button';
 
 interface ProductProps {}
