+++ src/contexts/AlertContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+import { IAlertContext } from '../types/alert';
+
+export const AlertContext = createContext<IAlertContext>({
+  isAlertOpen: null,
+  alertData: null,
+  showAlert: null,
+  hideAlert: null,
+});
