+++ src/types/alert.d.ts
@@ -1,0 +1,13 @@
+export interface IAlertData {
+  header: string;
+  subtitle: string | null;
+  message: string | null;
+  buttons: string[];
+}
+
+export interface IAlertContext {
+  isAlertOpen: boolean;
+  alertData: IAlertData;
+  showAlert: (IAlertData) => void;
+  hideAlert: () => void;
+}
